requirements.txt
alembic==1.0.5
argh==0.26.2
autopep8==1.4.3
Click==7.0
coverage==4.5.2
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
pathtools==0.1.2
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.14
visitor==0.1.3
watchdog==0.9.0
Werkzeug==0.14.1
WTForms==2.2.1
